chore: move to tests dir
