+++ tests/mocks/server.ts
@@ -1,9 +1,12 @@
 import express from "express";
 import asyncHandler from "express-async-handler";
-import { RelyingParty } from "../siopv2/siop";
+import { RelyingParty } from "../../src/siopv2/siop";
+import { Server } from "http";
+import { rp } from "./openid";
 
 const app = express();
 app.use(express.json());
+let server: Server;
 
 export const requestsMap = new Map<string, string>();
 
@@ -13,12 +16,16 @@ app.route("/siop/:id").get(
     })
 );
 
-export function startServer(rp: RelyingParty, port = 5000) {
+export function startServer(port = 5000) {
     app.route("/api/auth").post(
         asyncHandler(async (req, res) => {
             await rp.verifyAuthResponse(req.body);
             res.status(204).send();
         })
     );
-    app.listen(port);
+    server = app.listen(port);
+}
+
+export function stopServer() {
+    server.close();
 }
